requirements.txt
streamlit>=1.28.0
openai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
ollama>=0.1.0 